refactor : ProfileImage 폴더 이동 style -> components
